feat(Persistence): bring back the Migration from V1 code
