+++ MixTeam/Dependencies/Persistence/PersistenceError.swift
@@ -1,0 +1,3 @@
+enum PersistenceError: Error {
+    case cannotGetDocumentDirectoryWithUserDomainMask
+}
